Backend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-channels==0.7.0
django-cors-headers==3.14.0
django-leaflet==0.28.3
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-decouple==3.8
pytz==2023.3
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
